Separating Comms, Control, and adds Utillity Files
