+++ comms/rocketClass.hpp
@@ -9,7 +9,7 @@
 #ifndef _ROCKET_HPP_
 #define _ROCKET_HPP_
 #define numOfCParams 3
-#define commsDevice 123
+#define commsDevice 19
 #define fpacc 5
 #define numBytes 64
