+++ comms/rocketClassDef.cpp
@@ -80,39 +80,7 @@ float rocket::getRollRate(){
 }
 
 
-/*Converting a char aray to float (Found this online, dont know how well it works)*/
-float catof(char* num){
-    //Serial.println(num);
-    if (!num || !*num) return 0;
-    float rhs = 0;
-    float lhs = 0;
-    int divisor = 1;
-    int sign = 1;
-    bool inFraction = false;
-    if (*num == '-'){ ++num; sign = -1; }
-    else if (*num == '+'){ ++num; }
-    while (*num != '\0'){
-        if (isDigit(*num)){
-            if (inFraction){
-                divisor *= 10;
-                lhs += ((float)(*num - '0'))/divisor;
-            }
-            else{ 
-                rhs = rhs*10 + (*num - '0');
-            }
-        }
-        else if (*num == '.'){
-            if (inFraction)
-                return sign * (rhs + lhs);
-            else 
-                inFraction = true;
-        }
-        else 
-            return sign * (rhs + lhs);
-        ++num;
-    }
-    return sign * (rhs + lhs);
-}
+
 
 int rocket::fillModel(int fpsize, int devName){
     int property = 0;
@@ -129,30 +97,15 @@ int rocket::fillModel(int fpsize, int devName){
             case 1: moi = catof(str); break;
             case 2: plan.parseFlightPlan(str); break;
         }
-        delete[] str;
-        str = nullptr;
+        {
+          delete[] str;
+          str = nullptr;
+        }
         ++property;
     }
-    Serial.println(omega);
-    Serial.println(moi);
     return 0;
 }
 
 int rocket::logData(char* fname, int floatSize){
-    Serial.println(F("IN LOGDATA"));
-    //File file = SD.open(fname, FILE_WRITE);
-    Serial.println(F("OPENED FILE?"));
-    //if (file){
-        Serial.println(F("QUATERNION DATA"));
-        Serial.println(pftoi(Q[1]), floatSize);
-        Serial.println(pftoi(Q[2]), floatSize);
-        Serial.println(pftoi(Q[3]), floatSize);
-        Serial.println(pftoi(Q[4]), floatSize);
-        Serial.println(F("ALTITUDE"));
-        Serial.println(pftoi(z), floatSize);
-    //}
-    //else {return -1;}
-    //file.close();
-    Serial.println(F("OUT OF LOGDATA"));
-    return 0;
+    
 }
