+++ control/control.ino
@@ -10,6 +10,7 @@ Adafruit_BNO055 orient = Adafruit_BNO055(55);
 Servo ailerons;
 bool wireFlag = false;
 
+
 void setup() {
     //serialDump();
     pinMode(commsRst, OUTPUT);
