+++ control/flightplan.cpp
@@ -1,99 +1,5 @@
 #include "flightplan.hpp"
 
-/// Utility Functions
-
-bool isDigit(char c)
-{
-    return '0' <= c && c <= '9';
-}
-
-bool areDigits(char* c, int n)
-{
-    for (int i = 0; i < n; ++i)
-    {
-        if (!isDigit(c[i]))
-        {
-            return false;
-        }
-    }
-    return true;
-}
-
-int appendCharDigit(int number, char c)
-{
-    if (isDigit(c))
-    {
-        number *= 10;
-        number += c - '0';
-    }
-    return number;
-}
-
-int getNumberFromChars(char* c, int n)
-{
-    int result = 0;
-    for (int i = 0; i < n; ++i)
-    {
-        result = appendCharDigit(result, c[i]);
-    }
-    return result;
-}
-
-// Utility Functions for Config Parsing
-
-bool isFpVital(char e){
-    if (isDigit(e)) return true;
-    if (e == '.') return true;
-    if(e == '+') return true;
-    if (e == '-') return true;
-    if (e == '~') return true;
-    if (e == '#') return true;
-    if (e == ',') return true;
-    return false;
-}
-
-int getCaSize(char* str){
-    int i = 0;
-    while(str[i] != '\0'){ ++i; }
-    ++i;
-    return i;
-}
-
-char* caAppend(char* in, char e){
-    /*Appends to a char* allocated on the heap*/
-    if (in == nullptr){
-        in = new char[2];
-        in[0] = e;
-        in[1] = '\0';
-        return in;
-    } 
-    else{
-        int size = getCaSize(in);
-        char* temp = new char[size+2];
-        int i = 0;
-        for(; i < size-1; ++i){ temp[i] = in[i]; }
-        delete[] in;
-        in = temp;
-        in[i] = e;
-        in[i+1] = '\0';
-        return in;
-    }
-}
-
-int pftoi(float &f){
-    int a = f;
-    if (a != 0){
-        f -= a;
-        if (f < 0){f *= -1;}
-        return a;
-    }
-    else {
-        f *= 10;
-        a = f;
-        f -=a;
-        return a;
-    }
-}
 
 ///flightplan Implementation
